sort the dbs by title (closes #7)
Add the spaceship operator to the Database class - lets us use Ruby's
builtin Array#sort method to sort a list of dbs. Then call sort! on the
list of scanned databases in the scan_blast_db method, so that dbs are
always sorted.

Sorting dbs later (like in templates) might result in unexpected errors
as we rely on the db index value returned from the form to pick up the
right db for blasting.