gracefully handle lack of explicit path to bin directory (closes #4)
Not providing a config.yml, or an explicit path to the bin directory
would set SequenceServer.bin to be an empty string. This caused a
problem later as File.join would prepend a slash in front of the
executables causing them to be invalid.

Set SequenceServer.bin to be nil, in such cases, and use a helper
(executable()) that returns the correct path to the requried blast
binary.