properly decouple Helpers from App
A significant chang in scan_blast_db is to remove the depenedency on
App.binaries method. scan_blast_db now takes an additional parameter
whish is the path to blastdbcmd command. It makes sense to call
scan_blast_db from App.init this way passing to it the path to
blastdbcmd that the App has discovered.

Also the helpers should be given the absolute path of the directories to
scan for binaries, or database. Their only concern should be scanning
the given path and not interpolating (expanding) it.