dynamically derive configuration options
In 870aae27 I had defined parse_config which would set the value of each
configuration option. While this allowed for an explicit configuration
management throughout the app, the method parse_config in intself was
essentially inelegant, as it had to know in advance how to handle each
option.

The beauty of Sinatra::Base.set, and the elegance of Ruby's proc (yup,
the closure properties) makes it possible to derive configuration
options dynamically by defining by defining a proc for each setting
which knows exactly how to process its value.