raise error on empty sequence (about #14)
Sinatra returns an empty string for empty textboxes in the form.
Naturally 'unless sequence' would always evaluate to false. While it
fixes the original problem reported by NOLFXceptMe in #14, the error
handling still needs to improve ( which is currently handled by a poor
little 500.erb :P).

Perhaps, an ArgumentError should be reported to the user in the same
form, so that he can correct his mistake.