ui enhancements in the advanced parameters area
Use 'hand' cursor, and tooltips to indicate that "Advanced parameters"
is clickable. Clicking on the input box unfolds the list of advanced
parameters as before, but further clicks do not hide it, so that the
list does not disappear if the user reclicks the input box to
ensure/regain focus. Also, clicking on the list of advanced parameters
no longer hides it so that it is easier to copy paste from the list, or
just highlight some option. I felt that this would be a more natural
behavior.