fix ui bug with input sequence type detection
Earlier when a user pasted a sequence, we would trigger a sequence
detection, and accordingly disable incompatible blast methods, but we
never bothered if the user further edited the sequence, or typed in the
sequence, or emptied the sequence (cut, delete, etc). This commit
hopefully takes care of all that :).