ui logic rewrite (closes #8)
We now poll for user input change rather than trying to bind to all
possible events - keydown, cut, paste and what not. Polling captures it
all, so we define our UI behavior only once. This makes our code less
prone to errors, and much more manageable.

Another significant change is that we no longer reset the entire form in
one go. Instead, we only change the form just enough to reflect a change
in the user input. Advantage here is that if a user has already selected
a blast method, or databases to blast against, before he/she pastes the
input sequence, we don't reset the existing selection unless it is
incompatible with the input sequence type.

This is a major overhaul to our UI 'intelligence', made better over
various iterations, albeit a little heavy on the browser. This should
take good care of #8.