fetch BLAST results via AJAX on the onsubmit event and not onclick
`onsubmit` is triggered whenever the form is about to be submitted, regardless
of the input mode.  `onclick`, on the other hand, won't be triggered if the
form submission was triggered by pressing `Enter` key.

The rationale for this change is not only pragmatic, but also a bug that caused
the placeholder text for advanced parameters to be submitted on some old
browsers that don't support placeholders natively.  On such browsers,
SequenceServer mocks a placeholder by setting `value` attribute of the input
(text) box to the desired text, and removing it just before submitting the form
(if not already removed when filling the form).  However, fetching BLAST
results on `onclick` caused the form to be submitted before SequenceServer
could remove the mock placeholder text.

Thanks to Jonna Katharina Kulmuni, and Fourmidable for helping discover and
squash this issue.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>