ui: do not uncheck an appropriate blast method when sequence type changes (#83)
I don't see any reason why the line:

    $("#blastp, #tblastn").uncheck().disable().first().change();

would uncheck all the blast methods instead of just `blastp` and `tblastn`,
because of which SS had the bug on the first place.  Embarrassingly, I don't
even understand why this patch actually fixes the problem :|.  Thanks to Ben,
for pointing out the issue and the solution :).

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>