ui: provide an index for a multi query search (close #10)
Entries in the index corresponds to the results for each query.  Long query
description lines are capped at 22 characters, while the full length is shown
in the tooltip.

Of course, the index is not show if the query contains only one sequence.
Depending on whether SS needs to show the index or not, width of the results is
accordingly shrunk or grown.

The index is automatically created on the client side from the blast results
from a simple jQuery plugin (tailored to SS's needs).

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>