start testing routes
Instead of raising an ArgumentError and return "Internal Server Error (500)",
SS now returns "Bad Request (400)" (on lines of Twitter) to indicate an
incorrect request.  For testing, SS relies on minitest for rspec style
behavioral tests, and rack-test to test the routes.  The test coverage at the
moment is tiny, nevertheless it helped tighten request validation a bit.

This commit relates to #14 since it changes how an error in client request is
handled, and #69 since it lays the foundation for testing the routes.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>