do not guesstimate blast method for empty sequence
`guess_sequence_type` is now more explicit about whatever it has to say about
the sequence type, be it a heuristic or a probabilistic guess or not defined.
Previously it would return `undefined` for multiple cases causing confusion for
the caller.  Now `undefined` is returned only for empty sequences and the 'best
guess' for everything else.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>