ui: display advanced parameters in a modal
The user can request to see the allowed advanced parameters by clicking on the
"question mark icon" next to the input box for advanced parameters.

I borrowed the idea from how Github and Gmail display "keyboard shortcuts" when
you press "?".  Showing a list of advanced parameters that can be used with
SequenceServer is a much similar use case.

Also, a "question mark icon" is a widely acknowledged metaphor for "what is
this?" or "help!".

And displaying and hiding advanced parameters no longer interferes with the
index generated for a multi-query BLAST.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>