issue an error message if the browser doesn't have JavaScript
SS requires JavaScript to be present.  If the user wants to use web-blast on a
non-JS browser, I am sure there are other alternatives out there.  If someone
wants to send a pull request to make SS work again on a non-JS browser, I am
keen to discuss the options.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>