ui: remove the close button from the top of modal windows
Only one way of doing things should be less confusing to the users.  One may
argue that with two buttons a user can click on whichever is closer to the
pointer.  But in my experience users unconsciously reach out for the same
button all the time.  I do.  We keep the lower one, that says 'close' or 'got
it', because the user is expected to read the text -- we have small meaningful
messages anyway -- before she dismisses the modal dialog saying, "Fine. Thank
you."

Thanks to Yannick for the tip :).

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>