blast.rb: reliably capture stdout, stderr, and exit status of a process
The changes made in the last commit didn't work for blastn when called with
task option except when the task was 'megablast'.  And I have no friggin idea
why.  A simple workaround is to employ bash's I/O redirection: redirect stdout
and stderr of BLAST+ to two, different, known files, and read the content of
those files later on, when BLAST+ quits.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>