ui: make alternate blast method selection from dropdown 'usable' (#72)
When SS can't uniquely determine a BLAST method to use for the given input, it
sets the blast button to the one it thinks is most likely, and presents
possible options in a dropdown list, generated dynamically as the user input
changes, next to the BLAST button.  Now if the dropdown list was generated more
than once before a page refresh, selecting an alternate BLAST method from the
dropdown list might or might not have worked.  And whether it worked or not,
the BLAST button would jiggle more than once.  The cause?  I was attaching to
the click event a function to swap BLAST methods _every time_ the list was
generated.  So when the user clicked on an alternate BLAST method from the
dropdown list, the same event handler was executed multiple times, sometimes
un-swapping BLAST methods and jiggling the BLAST button each time.  Solution?
Don't bind to dynamically created elements; use event bubbling.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>