ui: do not guess sequence type for queries less than 10 characters (#72)
This was the original behavior.  But to make BLAST method autoselection easier,
I had changed this behavior a few commits back so that SS would _literally_
assume ultrashort queries to be protein.  But now that SS notifies user of the
sequence type detected, it needs to be a little more vigilant.  Telling a user
that 'aa' is a protein sequence is kind of embarrassing.  Hence reverting SS
back to the old behavior -- do not guess sequence type if the length of the
query is less than 10 characters.  So how does SS autoselect BLAST method when
the sequence type is unknown.  Short sequences are _most likely_ protein.
Based on this assumption, and the database type selected SS pre-selects the
BLAST method it thinks is the most appropriate.  And presents a list of
alternatives in a drop down menu next to the BLAST button.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>