Reset location.hash when user submits a BLAST request (close #34).
Once the BLAST results are available users shouldn't have to scroll down
themselves to see the result.  SS does this by setting location.hash to
'result' once the BLAST results have been fetched causing browser to jump
to the BLAST results.

Now, if users need to re-blast, they will have to go back to the search form,
edit parameters, and click BLAST.  There are three ways a user can go back to
the search form: scroll up, refresh page, hit back button.  `location.hash` is
retained in the first two approach.  And browsers don't react if one tries to
set `location.hash` to the same value that it holds.  So SS wouldn't jump to
results for the second time if the user scrolled up, or did a page refresh to
get back to the search form, after viewing the results.  Instead SS would stay
at the current scroll location.

As a workaround, SS now resets `location.hash` to '' when user clicks BLAST.
And sets it back to 'result', as usual, when BLAST results are available.
Effectively, SS should always _jump_ to results after a successful BLAST.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>