Depend on Sinatra 1.3.x.
I don't really know if SequenceServer breaks b/w Sinatra 1.2.0 and 1.3.2.
But since I refer to the source code of the latest Sinatra release during
development, it is probably best to depend on the same.

We depend on 1.3.x since Sinatra follows Semantic Versioning [1], so public API
won't change in a point release.

[1]: http://semver.org/

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>