Depend on Bundler 1.2.x.
Not specifying a version number will cause RubyGems to settle for _any_ release
of Bundler that it might find on the user's system.  But the locally installed,
older release of Bundler might not be API compatible with newer release that
SequenceServer actually depends on, and might cause SequenceServer to break on
the user's machine.

We need at least Bundler 1.0.13.  But since I refer to the code of latest
release, it is probably best to depend on same, lest I should introduce a
change that depends on the public API provided by the newest release.

We depend on Bundler 1.2.x assuming it follows Semantic Versioning [1], so the
public API won't change in a point release.

[1]: http://semver.org/

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>