Remove SequenceServer::App.binaries.
Now we just export NCBI BLAST+ binaries directory to PATH. It will work since
all command are executed from SequenceServer in a sub-shell.

How do we then ensure that NCBI BLAST+ is installed? When SequenceServer is
launched, it must run blastdbcmd to look for BLAST databases. If blastdbcmd
exists we assume other BLAST binaries to be present and conclude NCBI BLAST+ is
installed on the user's system. But if blastdbcmd can't be found, we tell the
user to install NCBI BLAST+ like before.

All in all, this simplifies the code to a good extent.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>