Make navbar scrollable with the rest of the page.
Fixed navbar means all URL fragment accessible elements on the page must have
padding-top = height of the navbar or a part of that element will hide behind
navbar. Fixed navbar also means adjustment to affix and scrollspy. Persistent
navbar doesn't add value to the end users either. So made it scrollable with
the rest of the page.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>