Validate num_threads setting.
Previously SequenceServer would assume 1 thread if num_threads wasn't set, or
set incorrectly in the config file. I think we should use a default value only
if one wasn't provided by the user. If the user explicitly sets a configuration
value, and it's incorrect, we should alert the user instead. Additionally, we
warn the user if num_threads is greater than 256.

Explicit validation caught an error in our code where num_threads, if not set
in config file, was assumed to be 0 in Ruby 1.8.7.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>