Add more tests to ensure correct alignment values
The tests ensure any change in the behaviour of qstart, qend, sstart,
and send is reflected according to any changes made in future BLAST.
There a slight inconsistency between blastn and other methods about
the way they are shown. blastn automatically swaps the start and end
coordinates according to frame values while for others it has to be
inferred.

Signed-off-by: Vivek Rai <vivekraiiitkgp@gmail.com>