Collapsible hits - change chevron style, use delegated event.
Chevron style now is black against white, and down and right. Older
white against black style was rather distracting, imo. And using
right instead of up makes the hit header look FASTA-ish when folded.

It must be noticed that fa-chevron-right class isn't used. Instead
fa-rotate-270 class is toggled on fa-chevron-down. This is because
fa-chevron-right and fa-chevron-down are of unequal width, and the
text in the header would shift on toggle.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>