Fix graph button problems, handle case when evalue is 0.
Graph buttons were behaving wrongly because of incorrect event binding.
Fix that. Also, all the hits were colored with same color when one of
the hits had evalue equal to 0. This is fixed by not considering the
hits with zero evalue so that the least positive value is the minimum
reference value for drawing graph.

Signed-off-by: Vivek Rai <vivekraiiitkgp@gmail.com>