Update BLAST::Report API. Fix NCBI link generator.
An HSP now holds a pointer to the Hit object it belongs to. Hit to Query. And
Query to Report. Leveraging this, we define Hit#querydb which simply forwards
to Report#querydb, move Report#which_blastdb to Hit#whichdb, and extend Hit
class with Links module instead of Report. This approach feels "just right".

Update link generators accordingly, fix NCBI link generator while at it (which
was broken because we had begun using Hit#accession instead of Hit#id several
commits back and thus couldn't infer if the hit was from nr database), and
simplify Report a bit.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>