Switch back to WEBrick.
We had switched to Thin with 1.0 release. But installing Thin requires
compiling native extensions on the user's machine. Which means working XCode
setup with CLI tools installed on Mac, and appropriate development tools and
headers on Linux. This goes against our mandate of "easy to setup, intuitive
to use". So we switch back to WEBrick. WEBrick being a part of stdlib just
works.

Thanks to @qmwugbt112 and Phil Rees
(https://groups.google.com/forum/#!topic/sequenceserver/Link1UCgXrQ) for
raising this issue.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>