Fix evalue bug.
evalues that weren't outputted by BLAST in scientific notation were rounded to
two decimal places, causing 0.001 to be shown as 0.00. Now we render in pretty
scientific notation all evalues less than 1 and greater than or equal to 10.

Rendering all evalues less than 1 in scientific notation avoids the above bug.
It's easier to compare 1.2e-1 to 2.6e-5 than 0.12 to 2.6e-5. Similarly, 1-100
is easier to compare in non-scientific notation. But because we are doing e-1,
for consistency it's probably a good idea to start with e+1 (10) on this side.
Hopefully we will never have to render evalues greater than or equal to 10.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>