Mostly finish up the interactive figure for the paper.
Tourist turned out to be buggy and not as flexible. So we turn to Shepherd,
which is prettier, not buggy, and more flexible. Shepherd 1.1.3 with Tether
1.1.0 (Shepherd's dependency) are used*.

NOTE:
shepherd.css is a combination of tether.min.css and shepherd-theme-arrows.css.
shepherd.min.js is combination of tether.min.js and shepherd.min.js.  tour.js
and tour.css is our tour code on top of Shepherd.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>