Render graphical overview from JSON.
Here, we make a React component for graphical overview, similar to
Kablammo, which adapts the data for t.js and passes it to t.js for
drawing the graph.

Additionally, we provide SVG and PNG and download functionality for
graphical overview and fix a bug that graphical overview would show
'View more' button for exactly 20 hits.

Signed-off-by: Alekhya Munagala <alekhya.m94@gmail.com>
Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>