Responsiveness tweaks.
Always show 'Team'. Show 'Demo server' and 'Source code' only for large
screens (>= 1200 px). Put a right border on .sidebar instead of a left
border on .main, so that the border is hidden with sidebar when the
sidebar is hidden. Collapse navbar for small (< 992 px) and extra
small devices (< 768 px) by updating Bootstrap CSS to 3.3.6 with
@grid-float-breakpoint set to @screen-md-min. Finally, make the
"hamburger icon" prominent.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>