Rename Poll#schedule to Poll#queue
The Pool#schedule method was a little misleading. We thought that the job is being scheduled to run at some specific time, but this isn't what happens. What happens is the job is queued to run as soon as server finishes up with previous jobs. So the Pool#queue method instead of Pool#schedule method makes more sense.