Added an error class and placed error checks to make error class robust.
Initially all the exception was handled by Job, which sometime lead to unknown
error messages. This made user receive a message unrelated to what actually
issue was, which was misleading for the users. With a dedicated Error class we
can accurately set checks to catch correct issue and therefore provide an
appropriate error message to the user.
Error checks in formatter.rb to catch error while report generations. Initially
formatter just assumed that it can generate report from archive files without
checking for errors, which lead to error in different places and misleading the
users.
Also added an error check for non-existing Job. [KeyError]