Prettify HSP in the browser; simplify DOM
Prettifying HSP in the browser lets us fill up all available horizontal
space. <table>s have been replaced with list of divs and pre tags. Less
markup is easier to read and navigate. Also add grapher like border and
padding around the HSPs to make it look more consistent with Kablammo.

Note:
We previously had one <tr> for each HSP, but now we will have as many
pre tags as number of lines (triplets) the HSP is split on. So longer
HSPs will add more nodes to the page. Hopefully this is not a problem.
I took this approach to flexibly control vertical spacing between the
lines using CSS. If this turns out to be a problem, we can change to
one <pre> for whole HSP and use <br/> for blank lines (as before).

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>