Update dockerfile
The current docker file simply installed the latest sequenceserver gem
to the container. This didn't work for beta releases which require --pre
switch to be used. Further, the setup required gems to be published to
RubyGems before pushing the corresponding tag to GitHub (otherwise the
automated build will end up creating a docker image with new tag but
with old version of SequenceServer. Clearly, this is not ideal. The best
way to remain in sync is to build docker image from the fresh git clone
provided by automated build. To this end, adapt Michał T. Lorenc's
dockerfile from #366 to use COPY instead of downloading tarball from
GitHub.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>