Ability to import BLAST XML
XML files can now be imported using the new `--import` flag. This
creates a fake job (fake because the job won't be run) and prints the
job id. The XML file is copied to the job directory so that if the user
accidentally deletes the original file, we don't have to worry about it.
Furthermore, we can provide a consistent experience with regards to the
job's lifetime.

API wise, the imported XML file is accessed for report generation using
by calling `BLAST::Job#imported_xml_file`. Also, a boolean property
(`imported_xml`) is set at top-level of the JSON sent to the server
which is used to toggle UI elements.

Caveat: Currently the user must manually visit localhost:4567/job_id to
view the report.

Signed-off-by: Iwo Pieniak <ivo.pieniak@gmail.com>
Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>