Add href='#' to download FASTA and alignment links in sidebar
This is required to make the links clickable by scripting. This is, for
example, required for testing.  To prevent default behaviour (ie, of
navigating to the said url) we make sure to return false from the
respective event handlers.

We don't use button tag instead because we have other links in the
sidebar that are simpler as an anchor tag, and bootstrap does not allow
mixing the two in a nav. To use buttons we will have to set margins and
paddings ourselves.

Signed-off-by: Anurag Priyam <a.priyam@icloud.com>