Downgrade to capybara 3.15
Capybara 3.16 onwards require Ruby 2.4. However, we want to keep
supporting Ruby 2.3 until next year because current Ruby 2.3 is the
default on the current Debian and Ubuntu LTS (supported until mid 2020).

Capybara was updated to the latest (3.28) in fa0d3d01. Turns out we get
rid of that deprecation warning with Capybara 3.15 as well.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>