Don't test on Ruby 2.4 on Travis
Most of our users are either going to have Ruby 2.3 or Ruby 2.5 (based
on the default Ruby version on current Mac, and in the current and the
next LTS versions of Debian and Ubuntu). Besides, the odds of the code
failing on 2.4 when it works on 2.3 and 2.5 is extremely low. So let's
cut down the runtime for us and not overuse a shared resource (Travis).

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>