Length distribution issue #409
The scale of the query and the hits is different when BLASTX is
selected. As the query is a translated nucleotide sequence, it should be
divided by 3 when presented with the hits which are proteins.

Following solution detects the algorithm in the constructor and applies
correct query.length making the visualisation correct.

Signed-off-by: Iwo Pieniak <ivo.pieniak@gmail.com>