Length distribution end of x-axis fix
The last tick of the x-scale was drawn incosistently in terms of spacing
and the size compared to the rest of the scale.

It was solved by supplying a new array of values that is used to create
identically-sized ticks + finishing the x-axis with the value of 105% of
the query value, what provides nice ending of the axis and sorts out the
problem.

Signed-off-by: Iwo Pieniak <ivo.pieniak@gmail.com>