X and Y axis outer tick fix
For the both axes we call the nice() function which rounds up the domain
to the nice value which is either 2,5, or multiples of 10. Effectively
it rounds up to the next major tick, which is what we desired from the
begining.

To remove the annoying outerTick from the Y-axis outerTickSize attribute
is set to 0 to simply remove it. In some situations, the last tick on
the Y-axis will be also the outer-tick as these can overlay. Apart from
this case, consistently the outer tick will not be present in the
y-axis.

Signed-off-by: Iwo Pieniak <ivo.pieniak@gmail.com>