Optimise rendering process
First update transfers responseJSON to Report.state so that summary
overview and circos can be rendered. Subsequent updates add Query, Hit,
and HSP components to Report.state.results which are then inserted into
the page by React. The first cycle is started with the call to setState
in setStateFromJSON and subsequent cycles are is queued in
componentDidUpdate. componentDidUpdate is run after the DOM has been
updated (that is, virtual DOM has been applied). The use of setTimeout
ensures that the browser gets a break from the rendering process and can
respond to user interactions. Because we queue the next update after DOM
has been updated, the setTimeout interval does not need to be very high.
Finally, the number of each HSPs rendered in a cycle should be small.
This is because the browser is blocked from responding to user
interactions such as click till it has finished the current cycle.
However, if this number is too small, like 1, the time required to
render the whole report increases. The implemented threshold of 3 is 3
times faster than rendering only 1 HSP in each update cycle.

The rendering process is further optimised by implementing the
shouldComponentUpdate life-cycle method for Query, Hit, and HSP
components. This avoids re-rendering (including to virtual-dom).
This causes the view sequence button per-hit to break as it
relies on Hit.state to activate SequenceViewer.

API wise, hits are no longer nested inside queries, and hsps are no
longer nested inside hits. This has created minor breaking changes
that should be addressed next.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>