Use domID to "link" hits and hsps
We were previously using hit number for this. However, hit number is not
unique across the page; it is reset for each query. Thus folding a hit
would fold other hits with the same number in other queries as well.

Using domID may feel convoluted but it is not. We should change the JSON
schema someday to be flat and have each result item (query, hit, and hsp
objects) keyed by their id, where the id would be constructed the same
was as domID.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>