Do not prompt to join mailing list when running under Docker
This is because SequenceServer creates a file in DOTDIR to remember
that it has asked the user once. However, this file doesn't persist
when running under Docker and SequenceServer tends to prompt users
repeatedly. Furthermore, a prompt can be annoying for automated
setups which is likely the primary use case of Docker build.

see #468

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>