Detect incompatible databases and offer to reformat them
This is as part of the same workflow that detects unformatted FASTAs
and offers to turn them to BLAST databases.

Database title is preserved when reformatting a BLAST database. If the
corresponding FASTA file does not exist, it is generated from the BLAST
database which is then formatted. Unfortunately taxonomy information is
not preserved.

This commit moves database formatting methods from Database class into
a separate module (MAKEBLASTDB). Accordingly, database_spec.rb has been
split into two files. This commit further removes the -u CLI option.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>