Display hit title in summary table of hits
The use of max-width 0 is based on https://stackoverflow.com/a/11877033
The fix for IE9 and lower suggested in that thread has not been applied:
I believe we do not support IE9 and lower. Other than that, I tested the
solution on latest Chrome, Firefox and Safari on macOS Catalina.

Feature request by Niek Art.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>