Use sequence id instead of accession for sequence retrieval
With V4 databases it was possible to use accession for retrieving
sequences. So we used accession. Thus freeing us from the need to
encode sequence ids which may contain the pipe special character.

However, with V5 databases, the entire sequence id must be used.
This is problematic if databases use sequence id of the format
'gnl|foo|bar' (#475), and potentially for other id formats too.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>