Check and prompt to reformat incompatible databases on startup
In 9c77c81c we made it possible to detect and reformat V4 databases.
However, the functionality had to be invoked using `-m` flag, meaning
users wouldn't be automatically cautioned if they had a mix of V4 and
V5 databases or databases created without the -parse_seqids option in
databases directory. We change that here.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>