Should quit if users do no format/reformat any database after viewing
That is, if users chose to view databases to format/reformat but did
not end up formatting/reformatting any, the expected behaviour is to
quit instead of redo.

Had accidentally changed the behaviour to redo in 9c77c81c. Realised
it when trying to make the behaviour of 'incompatible blast databases'
(introduced in the previous commit) same as that of 'no blast database
found'.