Revert to showing identity of top hit in the per-query hits table
Also change the title of the table back to 'Sequences producing
significant alignments' and of length distribution visualisation
to 'Length distribution of matching sequences'. This renaming is
because 'hit' can either mean the matching database sequence or
the matching region (HSP) to different users in different contexts.

Also, remove Hit#evalue and Hit#identity methods that are no longer
used and rename Hit#score to Hit#total_score for clarity.

close #506

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>