Remove the search history difference
We do not currently offer a mechanism to see past searches. It's on the
user to save and organise their searches, e.g., using their browser's
bookmarking features. And in future, we gain as much by making past
searches easily available to the user so that they do not use CPU
resources - the long-term-storage vs recompute trade-off lies with us.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>