Remove multiple instances options for pro and business tiers
We have not established a workflow for creating additional instances.
With one instance per tier, on the other hand, the workflow is clear
- user visits the website, selects tier, makes the purchase, fills
out the details, done - like with the first instance.

Furthermore, why would anyone want multiple instances for the same
set of colleagues? On fourmidable, the different instances are for
different groups of people - buy one for each. If the rationale is
clear organisation - we should spend time on improving our database
widget.

Finally, this removes the incentive for organisations to purchase a
single tier for ten labs. Instead, each lab makes their own purchase.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>