Sync capybara tests with inclusion of additional databases
Inclusion of additional databases to the spec/databases/v5 folder
(87dea4a6, 3f7dbfda) was causing capybara tests to fail.

Commit 13b2539f made matters worse by changing the database dir
for capybara tests, thinking that would fix the problem.

This commit reverts the change to database dir made in 13b2539f
and provides the correct fix for the failing test suite - which
was to include the new databases in one of the test cases.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>