Revise database format detection
Use %v option of "blastdbcmd -list_outfmt" to differentiate v4 and v5
databases and presence-absence of certain extensions to determine if
databases were created using -parse_seqids option or not. (credit:
Yannick)

Also, tweak the glob pattern so that we do not need to take database
type into account and can determine if -parse_seqids option was used
or not for multi-part databases as well. (close #557)

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>