Rename Config#merge to Hash#deep_merge
Config#merge was introduced in 2181d290 to deep merge Hashes.
In previous commit, there was a need to also define deep_copy
for Hashes, so we created refinements.rb to monkey patch Hash
class. Now we are moving Config#merge to where it should have
been defined in the first place.