Clear SequenceServer::Database.collection after each test context
Test suites may initialise SequenceServer using different databases
directories in spec/. Many of the databases in the different subdirs are
essentially the same (same content, same display name), just formatted a
little different: v4, v5, with or without -parse_seqids option. This can
result in an accumulation of databases with the same name in the process
and cause confusion. For example, Capybara test that require selecting a
database in the interface based on name may fail because of ambiguous
matches.

Automatically clearing SequenceServer::Database.collection fixes this
problem.