Try to make running Capybara tests on Travis more robust
Sometimes Capybara tests that download a file fail because they find
another file in place of the downloaded file. I think this happens
because clear_downloads method silently fails sometimes. As a result
we end up with multiple files in downloads directory and the tests
pick up the wrong file based on default glob order.

This commit includes makes two changes. First, a warning message is
printed if multiple files are found in the downloads directory when
a single file is expected. Second, the globbed files are sorted by
mtime so that hopefully the most recently downloaded file is picked.