Add explicit webrick dependency, update nokogiri
Webrick is required by the server, but is not declared as a gem
dependency. Newer systems might not have it installed by default
and on these systems the server won't start up.

Nokogiri (required for testing) was using an old verion that needs
compiling. Since then Nokogiri has distributed pre-compiled
packages that are easy and fast to install - bump the version.