Add a basic cloud sharing client
Cloud sharing is a new SequenceServer feature that allows sharing
search results pages with anyone. Since SequenceServer is a gem
that a lot of people run locally on their machines, sharing results
from it may be difficult. This feature allows uploading job results
to an instance hosted by SequenceServer Cloud and getting back
a shareable URL accessible from the public internet to those
who have a copy of the URL.