Replace flaky JS assertions with reliable Capybara methods
There's a few places in code where JS evals are used for assertions.

While they _sometimes_ work, they _sometimes_ also flake, because
of unfortunate timings. Update to use capybara's built-in matchers
that are smarter and can wait a bit for requests and animations to
before jumping to bad conclusions about the DOM state. There are a
few more of these in the codebase.