Slighly more graceful error handling when result downloads are unavailable
Show an HTML error page when FASTA downloads are unavailable due to
no access to the Sequence database. This can happen if the user has
move or deleted the database, or is looking at an imported job.

The latter scenario will also happen with the new Cloud sharing
feature, at least in the initial release.