Allow setting job lifetime to "Infinity"
When configuring an instance via config.yaml or comand line params
they are passed in as strings, so up until now there was no way to
configure job retention to infinite using these config methods.

Allow passing in "Infinity" for a job retention config.

Float::INFINITY can be kept for backwards compatibility as job_lifetime
param can be passed as Float::INFINITY in from ruby to the instance
initializer.