Record makeblastdb errors for debugging purposes
formatting databases seems to be happening in a forked process,
so printing the stderr does not seem to have the desired effect
to aid the debugging. The shellou utility already has a way to
store stderr as a separate file, so save it next to the dababase or
database directory that was being formatted when the error occured.